requirements.txt
pbr
python-glanceclient
storops
docopt
